requirements.txt
# Core Framework
flet>=0.19.0

# Computer Vision & Barcode
# Use opencv-python-headless for Raspberry Pi (lighter, no GUI dependencies)
opencv-python-headless>=4.8.0; platform_machine == "armv7l" or platform_machine == "aarch64"
opencv-python>=4.8.0; platform_machine != "armv7l" and platform_machine != "aarch64"
pyzbar>=0.1.9
numpy>=1.24.0

# OpenGL for GL context checking (optional but recommended)
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0; platform_machine != "armv7l" and platform_machine != "aarch64"

# API & Networking
requests>=2.31.0
aiofiles>=23.0.0
aiohttp>=3.9.0

# Shopping List Integration
python-bring-api>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database
# SQLite3 is included with Python

# Logging
loguru>=0.7.0

# System utilities
psutil>=5.9.0  # For system monitoring on Raspberry Pi

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality (optional, for development)
black>=23.0.0
mypy>=1.5.0
pylint>=3.0.0
flake8>=6.1.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
